// web/src/api/generated/models/daily-word-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 每日单词响应
 */
export interface DailyWordResponse {
    /**
     * ID
     */
    'id'?: number;
    /**
     * 单词
     */
    'word'?: string;
    /**
     * 美式音标
     */
    'pronunciationUs'?: string;
    /**
     * 英式音标
     */
    'pronunciationUk'?: string;
    /**
     * 美式发音音频URL
     */
    'audioUrlUs'?: string;
    /**
     * 英式发音音频URL
     */
    'audioUrlUk'?: string;
    /**
     * 翻译
     */
    'translation'?: string;
    /**
     * 例句
     */
    'example'?: string;
    /**
     * 例句翻译
     */
    'exampleTranslation'?: string;
    /**
     * 难度等级
     */
    'difficulty'?: number;
    /**
     * 难度描述
     */
    'difficultyDesc'?: string;
    /**
     * 单词分类
     */
    'category'?: string;
    /**
     * 单词笔记
     */
    'notes'?: string;
    /**
     * 发布日期
     */
    'publishDate'?: string;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
}